src/requirements-dev.txt
slixmpp~=1.8.5
mechanicalsoup~=1.3.0
ddt~=1.7.2
freezegun~=1.5.1
pytest-freezegun~=0.4.2
mock~=5.2.0
setuptools~=75.2.0
